+++ src/CHANGELOG.md
@@ -3,7 +3,8 @@
 v.4.6.2
 * [feature] All popups links built in frontend now respect the lang of Polylang plugin
 * [feature] Limiting the number of actors in the first popup page displayed (a click more link on the position defined in admin movie actors number)
-* [bug] Inside the popups, the links were not including the Polylang plugin lang
+* [technical] Better explained the purposed of language in advanced options, displaying the full language
+* [bug] Inside the popups, the links did not includ the Polylang plugin lang, so it wasn't translated
 
 v.4.6.1
 * [feature] Blocks translation, much work on blocks and javascripts for WP Classic edition